Measure a scorer at the granularities other work reports on, and fix the populations underneath
Detection accuracy on units drawn from inside a loop is the quantity this area
reports, and it is now measured rather than assumed. `unit_levels.py` cuts an
answer into sentences, blank-line chunks or the whole answer, reduces each unit
to one score, and reports area under the curve, average precision, accuracy and
balanced accuracy against a floor. The units match the granularities of the
nearest prior work; the operators do not reimplement anyone's detector, and the
module says where the correspondence stops.

Accuracy and area under the curve are saturated: twelve recipes differing in
selection rule, window, label family, horizon and class balance span 0.0014 in
area under the curve and 0.005 in accuracy, against a floor of 0.970 that a
scorer firing on nothing already reaches. Average precision is not saturated and
nobody reports it. Accuracy on a balanced subsample is inverted, ranking the
recipes with the least coverage of the run-up highest.

Four population faults are corrected, each of which moved a false-alarm rate:

Scores are stored at single precision. Half precision has a spacing of 5e-4 just
below one, so a score near the top of the range collapsed onto it, and ties at a
ceiling are what make a small budget unspendable. Storing at half precision
manufactured the failure the protocol reports.

Predictive entropy is averaged over a trailing window before it is inverted. Read
per token it reached its ceiling on 98.6% of healthy answers, because one
near-certain token is ordinary in healthy text. The width is the narrowest that
brings that share under a one percent budget.

A capped answer the judge could not rule on is excluded from every population
rather than counted as healthy. It was taking more than a tenth of the tightest
budget on answers that are probably positives. Answers the judge ruled not
degenerate stay negative, so the exclusion is keyed on the resolution and not on
whether an answer reached the cap. Evaluation refuses to guess which corpus a
directory describes when the scores were computed against another model's build.

Scoring silently inherits a filter belonging to training, which drops any rollout
the label family cannot label. For a token-level target that removes every answer
shorter than the labelling window: 846 healthy answers on validation, a quarter of
the class and all of them short. Evaluation now compares a scored table against
the split it claims to cover and records the shortfall beside the thresholds. The
scoring path itself is unchanged and still needs fixing.

The windowed repetition score gains the trailing form the comparisons need. The
stored one reads the window ahead of the token it labels, which is right for
annotating a finished answer and impossible for a live monitor, and the gap
between them is a factor of five in coverage of the run-up.

+++ scripts/score_rollouts.py
@@ -203,7 +203,7 @@ def score_split(
                     "num_tokens": int(length),
                     "onset_position": float(onset) if pd.notna(onset) else None,
                     "is_positive": bool(labels["is_positive"]) and pd.notna(onset),
-                    "scores": scores.astype(np.float16),
+                    "scores": scores.astype(np.float32),
                 }
             )
     return build_scores(records)
